PR Comments refactor
