+++ libs/client/home/src/lib/dashboard/canvas-list/canvas-card/canvas-skeleton.tsx
