+++ libs/server/canvas/src/lib/canvas.schema.ts
@@ -37,6 +37,13 @@ export class Canvas {
 
 export type CanvasDocument = Canvas & Document;
 export const CanvasSchema = SchemaFactory.createForClass(Canvas);
+
+@InputType()
+class CanvasInput {
+  @Field(() => String)
+  _id: CanvasId;
+}
+
 @InputType()
 export class CreateCanvasInput {
   @Field(() => String)
@@ -47,16 +54,19 @@ export class CreateCanvasInput {
 }
 
 @InputType()
-export class DeleteCanvasInput {
+export class StitchedCanvasInput extends CreateCanvasInput {
   @Field(() => String)
   _id: CanvasId;
 }
 
 @InputType()
-export class UpdateCanvasInput {
-  @Field(() => String)
-  _id: CanvasId;
+export class DeleteCanvasInput extends CanvasInput {}
 
+@InputType()
+export class GetCanvasInput extends CanvasInput {}
+
+@InputType()
+export class UpdateCanvasInput extends CanvasInput {
   @Field(() => String, { nullable: true })
   name?: string;
 
@@ -71,22 +81,4 @@ export class UpdateCanvasInput {
 
   @Field(() => Boolean, { nullable: true })
   isStitched?: boolean;
-}
-
-@InputType()
-export class GetCanvasInput {
-  @Field(() => String)
-  _id: CanvasId;
-}
-
-@InputType()
-export class StitchedCanvasInput {
-  @Field(() => String)
-  _id: CanvasId;
-
-  @Field(() => String)
-  name: string;
-
-  @Field(() => Boolean)
-  isPublic: boolean;
 }
